// mobile/src/api/generated/schemas/templateResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * GrowIt.API | v1
 * OpenAPI spec version: 1.0.0
 */
import type { TemplateExerciseResponse } from './templateExerciseResponse';

export interface TemplateResponse {
  id: string;
  name: string;
  /** @nullable */
  notes: string | null;
  exercises: TemplateExerciseResponse[];
}